Cast port number

+++ config/autoload/database.global.php
@@ -6,7 +6,7 @@ return [
     'db' => [
         'driver'   => getenv('DB_DRIVER'),
         'hostname' => getenv('DB_HOSTNAME'),
-        'port'     => getenv('DB_PORT'),
+        'port'     => (int)getenv('DB_PORT'),
         'database' => getenv('DB_DATABASE'),
         'user'     => getenv('DB_USER'),
         'password' => getenv('DB_PASSWORD'),
